+++ NEWS.md
@@ -5,3 +5,8 @@
 * Add euc_dist to calculate a raster of euclidian distance 
 # mdsFuncs 0.2.2
 * fixed bugs in mass splines 
+# mdsFuncs 0.2.3
+* added classify_quantile to classify raster by quantiles
+* added classify_equal_interval to classify raster by equal intervals
+
+